Centralize vehicle model formatting
